Started implementing PrimeJob class Fully implemented PrimeJob::Version Started implementing PrimeJob::Compression Code restructure in PrimesCpp with namespaces

+++ PrimesCpp/PrimesCpp.h
@@ -2,4 +2,6 @@
 
 #include <iostream>
 
-unsigned long long GetUlongInput();
+#include "ShortTypeNames.h"
+
+ulong GetUlongInput();
